requirements.txt
streamlit==1.39.0
transformers==4.46.0
torch==2.5.0
pillow==10.4.0
